Move some files out of the way while restructuring the app
