+++ streamlit-app/frontend/src/pages/inference.py
@@ -1,0 +1,4 @@
+import streamlit as st
+
+st.write("# Inference Time")
+
