+++ streamlit-app/frontend/src/reference/_scatter_plots.py
@@ -3,8 +3,8 @@ import pandas as pd
 import seaborn
 import streamlit as st
 
-from .plotting import with_large_text, task_type_colormap
-from .widgets import _add_persistent_selectbox, _add_axis_control
+from ._plotting import with_large_text, task_type_colormap
+from ._widgets import _add_persistent_selectbox, _add_axis_control
 
 
 def scatter_plot(data, x, y, hue: str | None = None, hue_order: list[str] | None = None):
